requirements.txt
openai
python-dotenv
syllapy
astral
requests
pytz
azure-functions
python-dateutil